dom/base/test/test_setInterval_from_start.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1252268
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1378394</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1378394">Mozilla Bug 1378394</a>

<script>
SimpleTest.waitForExplicitFinish();

// Some of our platforms are quite slow.  Use very large time values to
// try to avoid raciness.
const burnTimeMS = 5000;
const intervalTimeMS = 10000;

// The overall interval should include our callback "burn time".  So we
// expect the delay to be the remaining time.
const expectedDelayMS = intervalTimeMS - burnTimeMS;

// Allow some margin for error because of slow test platforms.
const allowedMarginMS = burnTimeMS / 2;

let id;
let lastEndTime;

function interval()
{
  let start = performance.now();

  if (lastEndTime !== undefined) {
    let delta = start - lastEndTime;
    ok(delta <= expectedDelayMS + allowedMarginMS,
       'interval should not fire too late');
    ok(delta >= expectedDelayMS - allowedMarginMS,
       'interval should not fire too early');
    clearInterval(id);
    SimpleTest.finish();
    return;
  }

  while((performance.now() - start) < burnTimeMS);

  lastEndTime = performance.now();
}

id = setInterval(interval, intervalTimeMS);
</script>

</body>
</html>
